edgeai.config: add ForceRegenRegistry option
:Release Notes:
add ForceRegenRegistry option to force the registry to be regenerated

:Detailed Notes:
Basically, the registry file is updated only when the webos_release
and webos_build_id are changed. However, there are cases where the
registry file needs to be updated even if the webos_release and
webos_build_id are not changed.

:Testing Performed:
Local test

:QA Notes:
N/A

:Issues Addressed:
N/A

Change-Id: I28a0ba9739fd3d053613f8c6cafdcbceb300c3db
Reviewed-on: http://gpro.lge.com/c/webosose/edge-ai-computer-vision/+/434876
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Junil Kim <jjunil79.kim@lge.com>
Reviewed-by: Junil Kim <jjunil79.kim@lge.com>
Static-Analysis: Junil Kim <jjunil79.kim@lge.com>